requirements.txt
Flask==2.0.3
requests==2.26.0
beautifulsoup4==4.10.0
lxml==4.6.4
tqdm==4.62.3
ffmpeg-python==0.2.0
python-dotenv==0.19.1